remove Makefile target install-info-debian, modify install-info to be compatible
* Makefile: remove target install-info-debian and modify target
  install-info so that it is compatible with the older dpkg version
  of install-info that resides in /usr/sbin and may be picked up
  when root is installing info files

Initial reporting by Jude DaShiell, suggestions for changing Makefile
by Nick Dokos.  The solution implemented re-arranges the arguments to
install-info to be compatible with both the dpkg and GNU version of
the program.  On Debian, /usr/bin/install-info is actually a wrapper
that calls /usr/bin/ginstall-info and issues the following warning
when called as root:

This is not dpkg install-info anymore, but GNU install-info
See the man page for ginstall-info for command line arguments

This warning can be safely ignored since we actually prefer that GNU
install-info is called even though we've arranged the arguments to
be compatible with dpkg install-info.

Tested on openSuSE Tumbleweed, Win7/Pro (both MSys and Cygwin) and by
Jude DaShiell on Debian Squeeze and Slackware.

TINYCHANGE