Improving bug reporting documentation
Hi Bastien and others,

Lately I have been seeing a lot of bug reports sent using the
org-submit-bug-report. The reporter usually reports the bug from a
session they have been using with their full blown customisations. And
often they are requested to reproduce the bug with emacs -Q. So I
thought it would be easier if the manual mentioned this little detail.
A patch is attached.

Hope this helps.

PS: I marked the patch as TINY CHANGE.