org-html.el: fix docstrings and the pre/postamble inserting.
* org-html.el (org-export-html-preamble)
(org-export-html-postamble): Fix docstrings.
(org-export-as-html): Insert the string used by a custom
function for `org-export-html-pre/postamble'.

The documentation is now in sync with the behavior: a custom
function for `org-export-html-pre/postamble' takes no argument,
and its value is inserted as a string.  Before that, users had
to use (insert ...) constructs to insert content in the preamble
through a custom function.

Thanks to Bill Jackson for spotting these inconsistencies.