Documentation: Clarified current role of ITEM special property
doc/org.texi (section on Special Properties) Clarified that ITEM refers
to the headline of the entry not the whole entry text.
(section on Matching Tags and Properties) Clarified that matching on
the value of ITEM special property is not currently supported,
and added a footnote about a partial workaround.
Also added missing -if suffix to references to org-agenda-skip-entry-if.

TINYCHANGE