org-mobile: Added five new targets for edit nodes
* lisp/org-mobile.el (org-mobile-edit): Added handling of
  addheading, refile, archive, archive-sibling and delete edit nodes.
(org-mobile-locate-entry): olp links containing only a file are now be
  located correctly.
(org-mobile-apply): Instead of finding the location of all target
  headings for edit nodes in a separate loop, they will be found
  immediately before applying edits.

org-mobile-apply needed to be changed, as the new edit nodes can
insert new headings or delete them, thereby changing the locations
of the target headings. Thanks to Aaron Peromsik for helping with
this patch.